refactor: split frontend api.ts (2079 lines) into api/ domain modules
- types.ts: all shared interfaces (types, settings, training, pending updates, log, backup, stats, airport, template)
- client.ts: API_URL, 3 axios instances (api/parserApi/hardwareApi), 401 interceptor, default export
- auth.ts: authApi
- parse.ts: parseApi (uses parserApi instance)
- flights.ts: flightsApi
- stats.ts: statsApi
- airports.ts: airportsApi
- achievements.ts: achievementsApi
- settings.ts: settingsApi
- analytics.ts: analyticsApi
- training.ts: trainingApi
- uploads.ts: uploadsApi
- setup.ts: setupApi
- admin.ts: adminApi
- notifications.ts: notificationsApi
- pendingUpdates.ts: pendingUpdatesApi
- backup.ts: backupApi
- template.ts: calculateDistance, templateApi, parserTemplatesApi
- index.ts: barrel re-exports all + default export for backward compat

0 TS errors, 128 frontend tests pass
