+++ frontend/src/lib/api/flights.ts
@@ -1,0 +1,60 @@
+import type {
+  Flight,
+  FlightFilters,
+  FlightInput,
+  FlightLookupResult,
+  GeoJSONFeatureCollection,
+} from "../../types";
+
+import { api } from "./client";
+
+// Flights API
+export const flightsApi = {
+  getAll: async (
+    filters?: FlightFilters
+  ): Promise<{
+    flights: Flight[];
+    total: number;
+    limit: number;
+    offset: number;
+  }> => {
+    const { data } = await api.get<{
+      flights: Flight[];
+      total: number;
+      limit: number;
+      offset: number;
+    }>("/flights", { params: filters });
+    return data;
+  },
+
+  getGeoJSON: async (filters?: FlightFilters): Promise<GeoJSONFeatureCollection> => {
+    const { data } = await api.get<GeoJSONFeatureCollection>("/flights/geo", {
+      params: filters,
+    });
+    return data;
+  },
+
+  getById: async (id: string): Promise<Flight> => {
+    const { data } = await api.get<Flight>(`/flights/${id}`);
+    return data;
+  },
+
+  create: async (flight: FlightInput, force = false): Promise<Flight> => {
+    const { data } = await api.post<Flight>(`/flights${force ? "?force=true" : ""}`, flight);
+    return data;
+  },
+
+  update: async (id: string, flight: Partial<FlightInput>): Promise<Flight> => {
+    const { data } = await api.put<Flight>(`/flights/${id}`, flight);
+    return data;
+  },
+
+  delete: async (id: string): Promise<void> => {
+    await api.delete(`/flights/${id}`);
+  },
+
+  lookup: async (params: { flightNumber: string; date?: string }): Promise<FlightLookupResult> => {
+    const { data } = await api.get<FlightLookupResult>("/flights/lookup", { params });
+    return data;
+  },
+};
