+++ frontend/src/lib/api/index.ts
@@ -1,0 +1,19 @@
+export * from "./types";
+export * from "./client";
+export { default } from "./client";
+export * from "./auth";
+export * from "./parse";
+export * from "./flights";
+export * from "./stats";
+export * from "./airports";
+export * from "./achievements";
+export * from "./settings";
+export * from "./analytics";
+export * from "./training";
+export * from "./uploads";
+export * from "./setup";
+export * from "./admin";
+export * from "./notifications";
+export * from "./pendingUpdates";
+export * from "./backup";
+export * from "./template";
